requirements.txt
pandas
seaborn
matplotlib
scipy
meteostat
requests
python-dotenv